// src/shared/ui/redesigned/Stack/model/types/typesStacks.ts
export type GridItems = 'start' | 'center' | 'end' | 'stretch';
export type GridContent = GridItems | 'between' | 'around' | 'evenly';
export type GridDirection = 'row' | 'column' | 'dense';

export type FlexJustify =
  | 'start'
  | 'center'
  | 'end'
  | 'between'
  | 'around'
  | 'unset';
export type FlexAlign = 'start' | 'center' | 'end' | 'stretch' | 'baseline';
export type ISeoTagName =
  | 'div'
  | 'main'
  | 'header'
  | 'footer'
  | 'section'
  | 'article'
  | 'aside'
  | 'nav';
export type FlexDirection = 'row' | 'column' | 'column-reverse';
export type FlexWrap = 'wrap' | 'nowrap' | 'wrap-reverse';

export type FlexGap = '4' | '8' | '16' | '24' | '32';